refactor: consolidate PDF test fixtures in obsidian-markdown/__tests__/
Move pdf-fixture.ts from mcp-core/__tests__/ to
obsidian-markdown/__tests__/ — the PDF parsers live in
obsidian-markdown/, so the fixtures belong alongside them. Consolidate
buildEmptyStreamPdf (was inlined in pdf.test.ts) into the shared
fixture module. Update all 4 import sites.
