+++ src/vault-mcp/mcp-core/__tests__/tool-definitions.test.ts
@@ -1591,7 +1591,8 @@ describe("file tool handlers", () => {
 
   it("returns structured markdown from a valid PDF", async () => {
     const { vault, readAsset } = await setupAssetHarness()
-    const { buildMinimalPdf } = await import("./pdf-fixture.js")
+    const { buildMinimalPdf } =
+      await import("../../obsidian-markdown/__tests__/pdf-fixture.js")
     await writeFile(join(vault, "doc.pdf"), buildMinimalPdf())
     const result = await readAsset({ path: "doc.pdf" })
     expect(result).toEqual({
